Move SierraDataGenerators into a new namespace
