+++ catalogue_pipeline/transformer/transformer_sierra/src/test/scala/uk/ac/wellcome/platform/transformer/sierra/generators/SierraDataGenerators.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.transformer.sierra.utils
+package uk.ac.wellcome.platform.transformer.sierra.generators
 
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
